+++ src/Elements/Generate/Geometry/Curve.g.cs
@@ -8,6 +8,7 @@ using Elements.GeoJSON;
 using Elements.Geometry;
 using Elements.Geometry.Solids;
 using Elements.Properties;
+using Elements.Validators;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -26,6 +27,12 @@ namespace Elements.Geometry
         [Newtonsoft.Json.JsonConstructor]
         public Curve()
         {
+            var validator = Validator.Instance.GetFirstValidatorForType<Curve>();
+            if(validator != null)
+            {
+                validator.Validate(new object[]{ });
+            }
+        
         }