+++ src/Elements/Generate/Geometry/Matrix.g.cs
@@ -8,6 +8,7 @@ using Elements.GeoJSON;
 using Elements.Geometry;
 using Elements.Geometry.Solids;
 using Elements.Properties;
+using Elements.Validators;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -25,6 +26,12 @@ namespace Elements.Geometry
         [Newtonsoft.Json.JsonConstructor]
         public Matrix(IList<double> @components)
         {
+            var validator = Validator.Instance.GetFirstValidatorForType<Matrix>();
+            if(validator != null)
+            {
+                validator.Validate(new object[]{ @components});
+            }
+        
             this.Components = @components;
         }