+++ src/Elements/Generate/Geometry/Transform.g.cs
@@ -8,6 +8,7 @@ using Elements.GeoJSON;
 using Elements.Geometry;
 using Elements.Geometry.Solids;
 using Elements.Properties;
+using Elements.Validators;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -25,6 +26,12 @@ namespace Elements.Geometry
         [Newtonsoft.Json.JsonConstructor]
         public Transform(Matrix @matrix)
         {
+            var validator = Validator.Instance.GetFirstValidatorForType<Transform>();
+            if(validator != null)
+            {
+                validator.Validate(new object[]{ @matrix});
+            }
+        
             this.Matrix = @matrix;
         }