+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - `Model` now has Elements property which is `IDictionary<Guid,Element>`.
 - `double Vector3.CCW(Vector3 a, Vector3 b, Vector3 c)`
 - `bool Line.Intersects(Line l)`
+- `Elements.Validators.IValidator` and the `Elements.Validators.Validator` singleton to provide argument validation during construction of user elements.
 
 ### Removed
 - The empty Dynamo project.